moved main logic of ExcelOptions to ExcelOptionsTrait
